requirements.txt
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.10
googledrivedownloader==0.4
idna==3.3
numpy==1.21.0
opencv-python==4.5.3.56
pandas==1.3.5
pdfminer.six==20200517
pdfplumber==0.5.28
pdfrw==0.4
Pillow==8.2.0
pyasn1==0.4.8
pycryptodome==3.10.1
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
python-dateutil==2.8.2
pytz==2021.3
pyzbar==0.1.8
qrcode==6.1
reportlab==3.5.67
requests==2.27.1
rsa==4.7.2
six==1.16.0
sortedcontainers==2.4.0
tk==0.1.0
urllib3==1.26.8
Wand==0.6.7
